+++ Kernel/Language/de.pm
@@ -1371,6 +1371,8 @@ sub Data {
         'Edit job' => 'Job bearbeiten',
         'Run job' => 'Job ausführen',
         'Affected Tickets' => 'Betroffene Tickets',
+        'Please remove the following words because they cannot be used for the ticket selection:'
+            => 'Bitten entfernen Sie die folgenden Wörter, da diese nicht für die Ticket-Selektion verwendet werden können:',
 
         # Template: AdminGenericInterfaceDebugger
         'GenericInterface Debugger for Web Service %s' => 'GenericInterface Debugger für Webservice',
@@ -2692,6 +2694,10 @@ sub Data {
         'Relative Period' => 'Relativer Zeitraum',
         'The last' => 'Die letzten',
         'Finish' => 'Abschließen',
+        'Please check restrictions of this stat for errors.'
+            => 'Bitte prüfen Sie die Beschränkungen dieser Statistik auf Fehler.',
+        'Please remove the following words because they cannot be used for the ticket restrictions:'
+            => 'Bitten entfernen Sie die folgenden Wörter, da diese nicht für die Ticket-Beschränkungen verwendet werden können:',
 
         # Template: AgentStatsEditSpecification
         'Permissions' => 'Rechtevergabe',
@@ -3081,6 +3087,8 @@ sub Data {
         'Template Name' => 'Name der Vorlage',
         'Pick a profile name' => 'Profilnamen auswählen',
         'Output to' => 'Ausgabe nach',
+        'Please remove the following words because they cannot be used for the search:' =>
+            'Bitte entfernen Sie die folgenden Wörter, da diese nicht für die Suche verwendet werden können:',
 
         # Template: CustomerTicketSearchResultShort
         'of' => 'von',
@@ -3132,6 +3140,8 @@ sub Data {
             'Auf diesem Bildschirm ist bereits ein Popup-Fenster geöffnet. Möchten Sie dieses schließen und dieses stattdessen öffnen?',
         'Please enter at least one search value or * to find anything.' =>
             'Bitte geben Sie zumindest einen Suchbegriff ein oder * um nach Allem zu suchen.',
+        'Please remove the following words from your search as they cannot be searched for:' =>
+            'Bitte entfernen Sie die folgenden Wörter aus Ihrer Suche, da diese nicht für die Suche verwendet werden können:',
         'Please check the fields marked as red for valid inputs.' => 'Bitte prüfen Sie die rot markierten Felder auf gültige Eingaben.',
         'Please perform a spell check on the the text first.' => 'Bitte führen Sie zuerst eine Rechtschreibprüfung aus.',
         'Slide the navigation bar' => 'Verschieben Sie die Navigationsleiste',