+++ Kernel/Modules/AdminGenericAgent.pm
@@ -248,6 +248,16 @@ sub Run {
             $Errors{ProfileInvalid} = 'ServerError';
         }
 
+        # Check if ticket selection contains stop words
+        my %StopWordsServerErrors = $Self->_StopWordsServerErrorsGet(
+            From    => $GetParam{From},
+            To      => $GetParam{To},
+            Cc      => $GetParam{Cc},
+            Subject => $GetParam{Subject},
+            Body    => $GetParam{Body},
+        );
+        %Errors = ( %Errors, %StopWordsServerErrors );
+
         # if no errors occurred
         if ( !%Errors ) {
 
@@ -288,6 +298,7 @@ sub Run {
             %GetParam,
             %DynamicFieldValues,
             %Errors,
+            StopWordsAlreadyChecked => 1,
         );
 
         # generate search mask
@@ -310,6 +321,7 @@ sub Run {
 
         # generate search mask
         my $Output = $LayoutObject->Header( Title => 'Edit' );
+
         $Output .= $LayoutObject->NavigationBar();
         $Output .= $LayoutObject->Output(
             TemplateFile => 'AdminGenericAgent',
@@ -705,6 +717,18 @@ sub _MaskUpdate {
         SelectedID => $JobData{NewLockID},
     );
 
+    # Show server errors if ticket selection contains stop words
+    my %StopWordsServerErrors;
+    if ( !$Param{StopWordsAlreadyChecked} ) {
+        %StopWordsServerErrors = $Self->_StopWordsServerErrorsGet(
+            From    => $JobData{From},
+            To      => $JobData{To},
+            Cc      => $JobData{Cc},
+            Subject => $JobData{Subject},
+            Body    => $JobData{Body},
+        );
+    }
+
     # REMARK: we changed the wording "Send no notifications" to
     # "Send agent/customer notifications on changes" in frontend.
     # But the backend code is still the same (compatibility).
@@ -726,8 +750,9 @@ sub _MaskUpdate {
     $LayoutObject->Block(
         Name => 'Edit',
         Data => {
-            %Param,
             %JobData,
+            %Param,
+            %StopWordsServerErrors,
         },
     );
 
@@ -1310,6 +1335,51 @@ sub _MaskRun {
     # build footer
     $Output .= $LayoutObject->Footer();
     return $Output;
+}
+
+sub _StopWordsServerErrorsGet {
+    my ( $Self, %Param ) = @_;
+
+    if ( !%Param ) {
+        $Self->{LayoutObject}->FatalError( Message => "Got no values to check." );
+    }
+
+    my %StopWordsServerErrors;
+    if ( !$Self->{TicketObject}->SearchStringStopWordsUsageWarningActive() ) {
+        return %StopWordsServerErrors;
+    }
+
+    my %SearchStrings;
+
+    FIELD:
+    for my $Field ( sort keys %Param ) {
+        next FIELD if !defined $Param{$Field};
+        next FIELD if !length $Param{$Field};
+
+        $SearchStrings{$Field} = $Param{$Field};
+    }
+
+    if (%SearchStrings) {
+
+        my $StopWords = $Self->{TicketObject}->SearchStringStopWordsFind(
+            SearchStrings => \%SearchStrings
+        );
+
+        FIELD:
+        for my $Field ( sort keys %{$StopWords} ) {
+            next FIELD if !defined $StopWords->{$Field};
+            next FIELD if ref $StopWords->{$Field} ne 'ARRAY';
+            next FIELD if !@{ $StopWords->{$Field} };
+
+            $StopWordsServerErrors{ $Field . 'Invalid' }        = 'ServerError';
+            $StopWordsServerErrors{ $Field . 'InvalidTooltip' } = $Self->{LayoutObject}->{LanguageObject}
+                ->Translate('Please remove the following words because they cannot be used for the ticket selection:')
+                . ' '
+                . join( ',', sort @{ $StopWords->{$Field} } );
+        }
+    }
+
+    return %StopWordsServerErrors;
 }
 
 1;