+++ CHANGES.md
@@ -23,6 +23,7 @@
  - 2015-01-06 Added ACLs handling for Owner and Responsible on screen load for AgentTicketOwner and AgentTicketResponsible.
 
 #4.0.8 2015-??-??
+ - 2015-04-23 Fixed bug#[11139](http://bugs.otrs.org/show_bug.cgi?id=11139) - StaticDB - Search incorrect.
  - 2015-04-21 Activated customer company support for the default customer user configuration.
  - 2015-04-20 Fixed bug#[10961](http://bugs.otrs.org/show_bug.cgi?id=10961) - Adding email recipients via addressbook does not update customer information.
  - 2015-04-17 Fixed bug#[11244](http://bugs.otrs.org/show_bug.cgi?id=11244) - Slave database not used in OTRS 4.